+++ tests/testthat/test-qenv_get_code.R
@@ -2,7 +2,7 @@ testthat::test_that("get_code returns code (character by default) of qenv object
   q <- qenv() |>
     eval_code(quote(x <- 1)) |>
     eval_code(quote(y <- x))
-  testthat::expect_equal(get_code(q), c("x <- 1", "y <- x"))
+  testthat::expect_equal(get_code(q), paste(c("x <- 1", "y <- x"), collapse = "\n"))
 })
 
 testthat::test_that("get_code returns code elements being code-blocks as character(1)", {
@@ -15,7 +15,7 @@ testthat::test_that("get_code returns code elements being code-blocks as charact
       z <- 5
     })
   )
-  testthat::expect_equal(get_code(q), c("x <- 1", "y <- x\nz <- 5"))
+  testthat::expect_equal(get_code(q), paste(c("x <- 1", "y <- x\nz <- 5"), collapse = "\n"))
 })
 
 testthat::test_that("get_code returns expression of qenv object if deparse = FALSE", {
@@ -24,7 +24,7 @@ testthat::test_that("get_code returns expression of qenv object if deparse = FAL
   q <- eval_code(q, quote(y <- x))
   testthat::expect_equivalent(
     toString(get_code(q, deparse = FALSE)),
-    toString(parse(text = q@code, keep.source = TRUE))
+    toString(parse(text = paste(c("{", q@code, "}"), collapse = "\n"), keep.source = TRUE))
   )
 })